refactor: change database & migration

+++ database/migrations/2024_04_03_122549_create_incomings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('incomings', function (Blueprint $table) {
             $table->id();
             $table->dateTimeTz('input_date');
-            $table->integer('delivery_id');
+            $table->unsignedBigInteger('delivery_id');
             $table->string('product_code');
             $table->timestamps();
